backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
geoalchemy2==0.14.2
pydantic>=2.7.0
pydantic-settings==2.3.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
opencage==2.2.0
shapely==2.0.2
pykml==0.2.0
httpx==0.25.1
python-dotenv==1.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
asyncpg==0.29.0
alembic==1.13.1
numpy<2
email-validator
bcrypt==3.2.2